add nghyperopt + improve test
